+++ requirements.txt
@@ -1,3 +1,3 @@
-ops==2.19.4
+ops==2.20.0
 requests==2.32.3
 cosl==0.0.57
