+++ tox.ini
@@ -111,4 +111,4 @@ deps =
     pytest-operator
     -r{toxinidir}/requirements.txt
 commands =
-    pytest -v --tb native --ignore={[vars]tst_path}unit --log-cli-level=INFO -s {posargs}
+    pytest -v --tb native --ignore={[vars]tst_path}unit --ignore={[var]tst_path}unit_harness --log-cli-level=INFO -s {posargs}
